Remove iskeyword< from b:undo_ftplugin and add path<
Possibly addresses https://github.com/guns/vim-clojure-static/issues/5

We moved the iskeyword definition to the syntax file, since it
is pertinent to syntax matching, but did not remove it from
b:undo_ftplugin.

I wasn't able to reproduce the above problem, but this certainly seems
like the culprit.

We also reset path< if it was set.