Only exceptional special form keywords are exceptions
Addresses #6, but not in the way the poster desires.

The purpose of highlighting language keywords is to provide affirmation
that the keyword exists and is spelled correctly.

The purpose of coloring keywords differently is to convey information
about their utility and limitations.

My colorscheme highlights macros as blue, so I am instantly reminded
when I see a blue word that I should not pass it as a higher-order
function. Similarly, if, def, and let are arguably the most important
primitive elements of a Lisp, and coloring them together with the
handful of other special forms convey information about their
performance and composability that would be lost when lumping them
together by their logical classification.

The Exception group is exceptional in that the utility of coloring 'try'
'catch' 'finally' and 'throw' differently to stand out from all other
keywords supersedes the fact they are special forms since it is commonly
expected that the exception-handling syntax of a language is itself
exceptional.