Rename clojurePattern to clojureRegexp
Java's regular expression class is called Pattern, so it is tempting
to name the syntax group the same way, but this name suffers from
generality and is also in conflict with Vim convention.

Grepping for 'syn\S*\s+\S+\s+\S*(regex|pattern)' in Vim's runtime
files shows that'langRegex(p)' is used extensively and exclusively for
regular expressions (naturally), while 'langPattern' is used solely by
syntax/sh.vim to refer to regular expressions.

(Additionally, Clojure refers to a #"" pattern as a "regex")

The name of this syntax group is significant because regular expressions
are essentially string-like in syntactical form, so plugins wishing
to match string-like sequences can match against the self-descriptive
pattern '\v\cstring|regex'.

This does not break any known plugins.