Do not match #(…) and #{…} separately
Closes #11

External rainbow parentheses scripts, like rainbow_parentheses.vim, work
by overriding the paren regions in syntax files. In order to remain
compatible with such scripts, any attempts to differentiate between
the different paren regions for `contains` scoping should be avoided.
Creation of non-standard paren regions also conflicts with such scripts.

Thus we remove the clojureAnonFn and clojureSet regions and allow the &
and % reader symbols in […] and (…) regions respectively, as well as the
top scope.

It should be mentioned that the putative reason for creating the
clojureAnonFn region was to more accurately contain % args, which is
only treated specially by the reader within a reader #() function
literal. However, the previous implementation didn't work correctly,
causing % args to be matched only when in the top form of the #()
literal.