Avoid the use of :normal in the indentexpr
Docs state that the indentexpr must not contain side effects, but that
moving the cursor is okay because it will be reset. However, a closer
look reveals that :normal commands for moving the cursor have side
effects beyond just moving the cursor and thus do not behave properly.

The Vim documentation should be changed to make this more clear, but we
can do our part now.