Ensure proper highlighting of legal Clojure symbols
This patch fixes highlighting of Clojure symbols that involve characters
such as "'", "#", and ":".

Previously a symbol like `super:hero` would incorrectly treat the
":hero" portion of the symbol as a clojureKeyword. Also if a symbol
contained a "'" or "#" character it would be highlighted. In addition to
this observation, if the symbol contained a keyword following the "'" or
"#" character, it would be highlighted. So the symbol `burger'time`
would produce the synIDs clojureSexp, clojureQuote, and clojureMacro
respectively.