Only allow leading 0 in numbers for octal literals
This is more restrictive than what the reader actually allows. The
following are all legal, and are evaluated as base-10:

        090M
        09e9
        09.9
        09/08

Only 090 and 090N literals are errors.

It's bad form to be writing literal numbers with leading zeroes unless
you mean it to be octal, so we won't condone their use.