Aggregate property names into #'character-properties
We now derive the majority of character property names and aliases from
private fields.

This would be unacceptable in any normal program, but seems justifiable
in our case, since we are only trying to occasionally procure valid
regular expression tokens.

If this becomes unstable in future Java releases, we can go back to
parsing the Unicode specs and filtering the list by failure to raise
exceptions when used in a Pattern.