Allow for the syntax to be included in another
It is hard to include the Clojure syntax in another (for example, code
blocks in Markdown) because some regions use contains=TOP, which does
not work if there are other syntax rules.  Putting all top-level
elements in a cluster solves the problem.

Since some rules (like clojureKeyword or many defined with :syntax
keyword) are dependent on the 'iskeyword' option, it is modified locally
in the syntax file, as recommended in the Vim manual, section 44.2.