Move 'lispwords' setting to ftplugin
The scheme filetype sets lispwords in its ftplugin, so we should set
lispwords in our ftplugin for compatibility.

This also prepares vim-clojure-static for the lispwords-global-local
patch I will be sending upstream as `setlocal lispwords=` effectively
resets lispwords to the global value instead of blanking it.

cf. issue #43

Finally, the new setlocal syntax is compact, which Bram always
appreciates.